Create UserMetric for data from app history load Module for common UserMetric and Metric methods Basic first metric: news users
