+++ app/views/layouts/_navigation_links.html.erb
@@ -3,3 +3,4 @@
 <li><%= nav_link 'Recurring', recurring_path %></li>
 <li><%= nav_link 'One-Time', onetime_path %></li>
 <li><%= nav_link 'Affiliate', affiliate_path %></li>
+<li><%= nav_link 'Users', users_path %></li>
