+++ config/routes.rb
@@ -7,6 +7,7 @@ Rails.application.routes.draw do
   get   'recurring' => 'home#recurring'
   get   'onetime' => 'home#onetime'
   get   'affiliate' => 'home#affiliate'
+  get   'users' => 'home#users'
   get   'chart_data' => 'home#chart_data'
 
   get   'reset_metrics' => 'home#reset_metrics'
