+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170418115800) do
+ActiveRecord::Schema.define(version: 20170418133200) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -53,6 +53,15 @@ ActiveRecord::Schema.define(version: 20170418115800) do
     t.text     "app_title"
     t.text     "shop"
     t.decimal  "revenue",      precision: 8, scale: 2
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
+
+  create_table "user_metrics", force: true do |t|
+    t.date     "metric_date"
+    t.integer  "number_of_installs"
+    t.integer  "number_of_uninstalls"
+    t.integer  "new_users"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
